gitignore

+++ src/util/StrengthComparator.java
@@ -7,7 +7,7 @@ import model.Card;
 public class StrengthComparator implements Comparator<Card> {
 	
 	public int compare(Card a, Card b) {
-		float aStrength = a.getStrength(), bStrength = b.getStrength();
+		double aStrength = a.getStrength(), bStrength = b.getStrength();
 		if (aStrength == bStrength) return 0;
 		return aStrength < bStrength ? -1 : 1;
 	}
